refactor: copilot review
