+++ src/routes/volume/detail/CreateRecurringJob.js
@@ -2,6 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { Form, Input, Icon, Button, Select, Checkbox, InputNumber, Tabs, Tooltip, Radio } from 'antd'
 import { ModalBlur, ReactCron } from '../../../components'
+import { parseRetainAge } from '../../../utils/recurringJob'
 
 const Option = Select.Option
 const FormItem = Form.Item
@@ -48,17 +49,6 @@ const noRetain = (val) => {
 
 // Tasks that support the age-based retention policy (system-backup isn't offered here).
 const AGE_BASED_TASKS = ['backup', 'snapshot']
-
-// Parse a Go duration string (e.g. "2h30m", "8760h", "10m") into hours + minutes.
-const parseRetainAge = (str) => {
-  if (!str) {
-    return { hours: 0, minutes: 0 }
-  }
-  const h = /(\d+)h/.exec(str)
-  const m = /(\d+)m/.exec(str)
-  const total = (h ? parseInt(h[1], 10) * 60 : 0) + (m ? parseInt(m[1], 10) : 0)
-  return { hours: Math.floor(total / 60), minutes: total % 60 }
-}
 
 const modal = ({
   item,
